requirements.txt
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.6
pydantic==2.5.2
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
typing_extensions==4.8.0
uvicorn==0.24.0
watchfiles==0.21.0
websockets==12.0
mysql-connector-python==8.2.0
requests==2.31.0
sqlalchemy==2.0.23
dataclasses-json==0.6.3
pymongo==4.6.1
phpserialize==1.3